@@ -1,131 +1,108 @@
 {
-  "name": "chptr",
-  "description": "Command Line tool to handle separate Markdown files with special markup notations and export as a complete document.",
-  "version": "0.3.6",
-  "author": {
-    "name": "Mathieu Leblanc",
-    "email": "spiky.ing@gmail.com",
-    "url": "https://github.com/spikying"
-  },
-  "publisher": "spikying",
+  "author": "Mathieu Leblanc",
   "bin": {
-    "chptr": "./bin/run"
+    "chptr": "./bin/run.js"
   },
-  "bugs": "https://github.com/spikying/chptr/issues",
   "dependencies": {
-    "@braintree/sanitize-url": "6.0.0",
-    "@oclif/command": "1.7.0",
-    "@oclif/config": "1.16.0",
-    "@oclif/errors": "1.3.3",
-    "@oclif/plugin-help": "3.1.0",
-    "@oclif/plugin-not-found": "1.2.4",
-    "@oclif/plugin-plugins": "1.9.0",
-    "@types/convict": "5.2.1",
-    "@types/debug": "4.1.5",
-    "@types/deep-diff": "1.0.0",
-    "@types/diff": "4.0.2",
-    "@types/glob": "7.1.3",
-    "@types/inquirer": "7.3.0",
-    "@types/js-yaml": "3.12.5",
-    "@types/json5": "0.0.30",
-    "@types/latinize": "0.2.15",
-    "@types/mkdirp": "1.0.1",
-    "@types/node-notifier": "6.0.1",
-    "@types/sanitize-filename": "1.6.3",
-    "@types/tmp": "0.2.0",
-    "@types/uuid": "8.0.0",
-    "@types/validator": "13.1.0",
-    "@types/yaml": "1.9.7",
-    "chalk": "4.1.0",
-    "cli-ux": "5.4.9",
-    "convict": "6.2.4",
-    "convict-format-with-validator": "6.0.0",
-    "debug": "4.1.1",
-    "deep-diff": "1.0.2",
-    "diff": "4.0.2",
-    "glob": "7.1.6",
-    "inquirer": "7.3.3",
-    "inquirer-fuzzy-path": "2.3.0",
-    "js-yaml": "3.14.0",
-    "json5": "2.2.2",
-    "latinize": "0.4.1",
-    "minimatch": "3.0.5",
-    "mkdirp": "1.0.4",
-    "moment": "2.29.4",
-    "node-notifier": "8.0.1",
-    "sanitize-filename": "1.6.3",
-    "simple-git": "3.15.0",
-    "tmp-promise": "3.0.2",
-    "tslib": "2.0.0",
-    "typescript-ioc": "3.2.2",
-    "validator": "13.7.0",
-    "word-regex": "0.1.2",
-    "yaml": "1.10.0"
+    "@oclif/core": "^3",
+    "@oclif/errors": "^1.3.6",
+    "@oclif/plugin-help": "^6",
+    "@oclif/plugin-plugins": "^4",
+    "convict": "^6.2.4",
+    "convict-format-with-validator": "^6.2.0",
+    "debug": "^4.3.4",
+    "deep-diff": "^1.0.2",
+    "diff": "^5.1.0",
+    "glob": "^10.3.10",
+    "inquirer": "^8.2.6",
+    "inquirer-fuzzy-path": "^2.3.0",
+    "js-yaml": "^4.1.0",
+    "latinize": "^1.0.0",
+    "moment": "^2.30.1",
+    "node-notifier": "^10.0.1",
+    "sanitize-filename": "^1.6.3",
+    "simple-git": "^3.22.0",
+    "tmp-promise": "^3.0.3",
+    "typescript-ioc": "^3.2.2",
+    "validator": "^13.11.0",
+    "word-regex": "^0.1.2",
+    "yaml": "^2.3.4"
   },
+  "description": "CLI tool to handle separate Markdown files with special markup notations and export as a complete document",
   "devDependencies": {
-    "@oclif/dev-cli": "1.22.2",
-    "@oclif/test": "1.2.6",
-    "@oclif/tslint": "3.1.1",
-    "@types/node": "14.0.27",
-    "globby": "11.0.1",
-    "lodash": ">=4.17.21",
-    "nyc": "15.1.0",
-    "pkg": "4.4.9",
-    "prettify": "0.1.7",
-    "rimraf": "3.0.2",
-    "ts-node": "8.10.2",
-    "tslint": "6.1.2",
-    "tslint-config-prettier": "1.18.0",
-    "typescript": "3.9.7"
+    "@oclif/prettier-config": "^0.2.1",
+    "@oclif/test": "^3",
+    "@types/chai": "^4",
+    "@types/convict": "^6.1.6",
+    "@types/debug": "^4.1.12",
+    "@types/deep-diff": "^1.0.5",
+    "@types/diff": "^5.0.9",
+    "@types/inquirer": "^9.0.7",
+    "@types/js-yaml": "^4.0.9",
+    "@types/latinize": "^0.2.18",
+    "@types/mocha": "^9.0.0",
+    "@types/node": "^18",
+    "@types/node-notifier": "^8.0.5",
+    "@types/tmp": "^0.2.6",
+    "@types/validator": "^13.11.7",
+    "chai": "^4",
+    "eslint": "^8.56.0",
+    "eslint-config-oclif": "^5",
+    "eslint-config-oclif-typescript": "^3",
+    "eslint-config-prettier": "^9.1.0",
+    "mocha": "^10",
+    "oclif": "^3.17.2",
+    "prettier": "^3.2.4",
+    "shx": "^0.3.3",
+    "ts-node": "^10.9.2",
+    "typescript": "^5"
   },
   "engines": {
-    "node": ">=8.0.0"
+    "node": ">=18.0.0"
   },
   "files": [
     "/bin",
-    "/lib",
-    "/npm-shrinkwrap.json",
-    "/oclif.manifest.json",
-    "/templates"
+    "/dist",
+    "/oclif.manifest.json"
   ],
   "homepage": "https://github.com/spikying/chptr",
-  "keywords": [
-    "oclif",
-    "chptr",
-    "writing",
-    "cli",
-    "pandoc",
-    "markdown"
-  ],
   "license": "MIT",
-  "main": "lib/index.js",
+  "main": "dist/index.js",
+  "name": "chptr",
   "oclif": {
-    "commands": "./lib/commands",
     "bin": "chptr",
+    "dirname": "chptr",
+    "commands": "./dist/commands",
     "plugins": [
       "@oclif/plugin-help",
-      "@oclif/plugin-not-found",
       "@oclif/plugin-plugins"
-    ]
+    ],
+    "topicSeparator": " ",
+    "topics": {
+      "hello": {
+        "description": "Say hello to the world and others"
+      }
+    }
   },
   "repository": "spikying/chptr",
   "scripts": {
-    "postpack": "rimraf oclif.manifest.json",
-    "prepack": "tsc -b -v && oclif-dev manifest && oclif-dev readme",
-    "test": "echo NO TEST",
-    "version": "oclif-dev readme && git add README.md",
-    "testadd": "DEBUG='command:add' ./bin/run.cmd add --path='../testNovel'",
-    "testantidote": "DEBUG='command:antidote' ./bin/run.cmd antidote --path='../testNovel'",
-    "testbuild": "DEBUG='command:build' ./bin/run.cmd build --path='../testNovel'",
-    "testdelete": "DEBUG='command:delete' ./bin/run.cmd delete --path='../testNovel'",
-    "testedit": "DEBUG='command:edit' ./bin/run.cmd edit --path='../testNovel'",
-    "testinit": "DEBUG='command:init' ./bin/run.cmd init --path='../testNovel'",
-    "testrename": "DEBUG='command:rename' ./bin/run.cmd rename --path='../testNovel'",
-    "testreorder": "DEBUG='command:reorder' ./bin/run.cmd reorder --path='../testNovel'",
-    "testsave": "DEBUG='command:save' ./bin/run.cmd save --path='../testNovel'",
-    "testsplit": "DEBUG='command:split' ./bin/run.cmd split --path='../testNovel'",
-    "testtrack": "DEBUG='command:track' ./bin/run.cmd track --path='../testNovel'",
-    "buildstandalone": "pkg ."
+    "build": "shx rm -rf dist && tsc -b",
+    "clean": "shx rm -rf dist && shx rm tsconfig.tsbuildinfo && shx rm oclif.manifest.json",
+    "format": "prettier **/*.ts --write",
+    "lint": "eslint . --ext .ts",
+    "lint-fix": "eslint . --fix --ext .ts",
+    "postpack": "shx rm -f oclif.manifest.json",
+    "posttest": "npm run lint",
+    "prepack": "npm run build && oclif manifest && oclif readme",
+    "prepare": "npm run build",
+    "run-dev": "npm run build && env DEBUG=* node ./bin/dev --help",
+    "test": "mocha --forbid-only \"test/**/*.test.ts\"",
+    "version": "oclif readme && git add README.md"
   },
-  "types": "lib/index.d.ts"
+  "version": "1.0.0",
+  "bugs": "https://github.com/spikying/chptr/issues",
+  "keywords": [
+    "oclif"
+  ],
+  "types": "dist/index.d.ts"
 }
