@@ -1,32 +1,34 @@
-import { cli } from 'cli-ux'
-import * as d from 'debug'
-import minimatch = require('minimatch')
+import { ux } from '@oclif/core'
 import * as moment from 'moment'
-import * as simplegit from 'simple-git/promise'
-import { LogOptions } from 'simple-git/promise'
-import { MoveSummary } from 'simple-git/typings/response'
-import { format } from 'url'
-import { Inject, Singleton } from 'typescript-ioc'
+import { LogOptions, MoveSummary, SimpleGit } from 'simple-git'
+import { Inject, InjectValue, Singleton } from 'typescript-ioc'
 
 import { ChapterId } from './chapter-id'
 import { SoftConfig } from './soft-config'
 
-const debug = d('git-utils')
+import minimatch = require('minimatch')
+
+const debug = require('debug')('git-utils')
 
 @Singleton
 export class GitUtils {
-  private readonly _git: simplegit.SimpleGit
-  private readonly softConfig: SoftConfig
+  private readonly _git: SimpleGit
   private _gitConfigAdded: boolean
+  private readonly softConfig: SoftConfig
 
-  constructor(softConfig: SoftConfig, rootPath: string) {
+  constructor(@Inject softConfig: SoftConfig, @InjectValue('git') git: SimpleGit) {
     debug('CONSTRUCTOR GIT-UTILS')
-    this._git = simplegit(rootPath)
+    this._git = git
     this._gitConfigAdded = false
     this.softConfig = softConfig
   }
 
-  //todo: split git.add and git.commit to enable doing everything without committing, e.g. for functions like rename or others that don't need a commit automatically
+  // todo: split git.add and git.commit to enable doing everything without committing, e.g. for functions like rename or others that don't need a commit automatically
+
+  public async add(files: string | string[]) {
+    const git = await this.git()
+    return git.add(files)
+  }
 
   public async CommitToGit(
     message: string,
@@ -39,7 +41,7 @@ export class GitUtils {
     toStageFiles = toStageFiles || (await this.GetGitListOfStageableFiles())
     if (toStageFiles.length > 0 || forDeletes) {
       // try {
-      cli.action.start('Saving file(s) in repository'.actionStartColor())
+      ux.action.start('Saving file(s) in repository'.actionStartColor())
 
       // await this.coreUtils.processChapterFilesBeforeSaving(toStageFiles)
       await preProcessingCallback(toStageFiles)
@@ -50,16 +52,14 @@ export class GitUtils {
       }
 
       const commitSummary = await git.commit(message)
-      const hasRemote: boolean = await git.getRemotes(false).then(result => {
-        return result.find(value => value.name === 'origin') !== undefined
-      })
+      const hasRemote: boolean = await git.getRemotes(false).then(result => result.find(value => value.name === 'origin') !== undefined)
       if (hasRemote) {
         await git.push()
         await git.pull()
       }
 
       const toStagePretty = toStageFiles.map(f => `\n    ${f}`.infoColor())
-      cli.action.stop(
+      ux.action.stop(
         `\nCommited and pushed ${commitSummary.commit.resultHighlighColor()}:\n${message.infoColor()}\nFile${
           toStageFiles.length > 1 ? 's' : ''
         }:${toStagePretty}`.actionStopColor()
@@ -70,6 +70,39 @@ export class GitUtils {
     }
   }
 
+  public async GetGitContentOfHistoryFile(hash: string, file: string): Promise<string> {
+    debug(`getting content of history file: ${file}`)
+    const git = await this.git()
+    const content = await git.show([`${hash}:${file}`])
+    return content
+  }
+
+  // public async checkIsRepo(): Promise<boolean> {
+  //   return git.checkIsRepo()
+  // }
+
+  public async GetGitListOfHistoryFiles(
+    sinceDays: number
+  ): Promise<{ chapterFiles: string[]; date: moment.Moment; hash: string; summaryFiles: string[] }[]> {
+    debug('In GetGitListOfHistoryFiles')
+    const allLastCommitsPerDay = await this.GetAllLastCommitsPerDay(sinceDays)
+    const value: { chapterFiles: string[]; date: moment.Moment; hash: string; summaryFiles: string[] }[] = []
+
+    for (const commit of allLastCommitsPerDay) {
+      const allFilesInCommit = await this.GetAllFilesInCommit(commit.hash)
+      const chapterFiles = allFilesInCommit.filter(
+        f => this.softConfig.chapterRegex(true).test(f) || this.softConfig.chapterRegex(false).test(f)
+      )
+      const summaryFiles = allFilesInCommit.filter(
+        f => this.softConfig.summaryRegex(true).test(f) || this.softConfig.summaryRegex(false).test(f)
+      )
+      value.push({ chapterFiles, date: moment(commit.date).startOf('day'), hash: commit.hash, summaryFiles })
+    }
+
+    debug(`value: ${JSON.stringify(value)}`)
+    return value
+  }
+
   public async GetGitListOfStageableFiles(chapterId?: ChapterId): Promise<string[]> {
     const git = await this.git()
     const gitStatus = await git.status()
@@ -78,6 +111,7 @@ export class GitUtils {
       if (!value) {
         return value
       }
+
       return value.replace(/"(.*)"/, '$1')
     }
 
@@ -97,13 +131,13 @@ export class GitUtils {
 
     return unfilteredFileList
       .filter(val => val !== '')
-      .filter(val => {
-        return chapterId
+      .filter(val =>
+        chapterId
           ? minimatch(val, this.softConfig.chapterWildcardWithNumber(chapterId)) ||
-              minimatch(val, this.softConfig.metadataWildcardWithNumber(chapterId)) ||
-              minimatch(val, this.softConfig.summaryWildcardWithNumber(chapterId))
+            minimatch(val, this.softConfig.metadataWildcardWithNumber(chapterId)) ||
+            minimatch(val, this.softConfig.summaryWildcardWithNumber(chapterId))
           : true
-      })
+      )
   }
 
   public async GetGitListOfUntrackedFiles(): Promise<string[]> {
@@ -114,15 +148,43 @@ export class GitUtils {
       if (!value) {
         return value
       }
+
       return value.replace(/"(.*)"/, '$1')
     }
 
     return gitStatus.not_added.map(unQuote).filter(val => val !== '')
   }
 
-  // public async checkIsRepo(): Promise<boolean> {
-  //   return git.checkIsRepo()
-  // }
+  public async GetGitListOfVersionsOfFile(
+    filepath: string,
+    extractAll: boolean
+  ): Promise<
+    {
+      content: string
+      date: moment.Moment
+      file: string
+      hash: string
+      subject: string
+    }[]
+  > {
+    const git = await this.git()
+    const file = this.softConfig.mapFileToBeRelativeToRootPath(filepath)
+    const beginBlock = '########'
+    const endFormattedBlock = '------------------------ >8 ------------------------'
+    const gitLogArgs = ['log', '-c', '--follow', `--pretty=format:"${beginBlock}%H;%aI;%s${endFormattedBlock}"`]
+    if (!extractAll) {
+      gitLogArgs.push(`--since="${moment().add(-1, 'week')}"`)
+    }
+
+    return (await git.raw([...gitLogArgs, file]))
+      .split(beginBlock)
+      .filter(l => l !== '')
+      .map(l => {
+        const s = l.split(endFormattedBlock)
+        const logArray = s[0].split(';')
+        return { content: s[1] || '', date: moment(logArray[1]), file, hash: logArray[0], subject: logArray[2] }
+      })
+  }
 
   public async mv(from: string | string[], to: string): Promise<MoveSummary> {
     const git = await this.git()
@@ -134,115 +196,54 @@ export class GitUtils {
     return git.rm(paths)
   }
 
-  public async add(files: string | string[]) {
-    const git = await this.git()
-    return git.add(files)
-  }
-
   public async showHeadVersionOfFile(filepath: string): Promise<string> {
     const git = await this.git()
     return git.show([`HEAD:${this.softConfig.mapFileToBeRelativeToRootPath(filepath).replace(/\\/, '/')}`])
-  }
-
-  public async GetGitListOfVersionsOfFile(
-    filepath: string,
-    extractAll: boolean
-  ): Promise<
-    {
-      file: string
-      hash: string
-      date: moment.Moment
-      subject: string
-      content: string
-    }[]
-  > {
-    const git = await this.git()
-    const file = this.softConfig.mapFileToBeRelativeToRootPath(filepath)
-    const beginBlock = '########'
-    const endFormattedBlock = '------------------------ >8 ------------------------'
-    const gitLogArgs = ['log', '-c', '--follow', `--pretty=format:"${beginBlock}%H;%aI;%s${endFormattedBlock}"`]
-    if (!extractAll) {
-      gitLogArgs.push(`--since="${moment().add(-1, 'week')}"`)
-    }
-    return (await git.raw([...gitLogArgs, file]))
-      .split(beginBlock)
-      .filter(l => l !== '')
-      .map(l => {
-        const s = l.split(endFormattedBlock)
-        const logArray = s[0].split(';')
-        return { file, hash: logArray[0], date: moment(logArray[1]), subject: logArray[2], content: s[1] || '' }
-      })
-  }
-
-  public async GetGitListOfHistoryFiles(
-    sinceDays: number
-  ): Promise<{ hash: string; date: moment.Moment; chapterFiles: string[]; summaryFiles: string[] }[]> {
-    debug('In GetGitListOfHistoryFiles')
-    const allLastCommitsPerDay = await this.GetAllLastCommitsPerDay(sinceDays)
-    const value: { hash: string; date: moment.Moment; chapterFiles: string[]; summaryFiles: string[] }[] = []
-
-    for (const commit of allLastCommitsPerDay) {
-      const allFilesInCommit = await this.GetAllFilesInCommit(commit.hash)
-      const chapterFiles = allFilesInCommit.filter(f => {
-        return this.softConfig.chapterRegex(true).test(f) || this.softConfig.chapterRegex(false).test(f)
-      })
-      const summaryFiles = allFilesInCommit.filter(f => {
-        return this.softConfig.summaryRegex(true).test(f) || this.softConfig.summaryRegex(false).test(f)
-      })
-      value.push({ hash: commit.hash, date: moment(commit.date).startOf('day'), chapterFiles, summaryFiles })
-    }
-    debug(`value: ${JSON.stringify(value)}`)
-    return value
-  }
-
-  public async GetGitContentOfHistoryFile(hash: string, file: string): Promise<string> {
-    debug(`getting content of history file: ${file}`)
-    const git = await this.git()
-    const content = await git.show([`${hash}:${file}`])
-    return content
   }
 
   private async GetAllFilesInCommit(hash: string): Promise<string[]> {
     const git = await this.git()
 
     const allFiles = await git.raw(['ls-tree', '-r', '--name-only', hash])
-    return allFiles.split('\n').filter(f => f)
+    return allFiles.split('\n').filter(Boolean)
+  }
+
+  private async GetAllLastCommitsPerDay(sinceDays: number): Promise<LogFields[]> {
+    const git = await this.git()
+
+    const options: LogOptions<LogFields> | any = {}
+    options.splitter = ';'
+    options.format = {
+      date: '%aI',
+      hash: '%h'
+    }
+    if (sinceDays > 0) {
+      options['--since'] = `"${moment().subtract(sinceDays, 'days')}"`
+    }
+
+    const allCommits = await git.log(options)
+    const allDates: string[] = []
+    const allLastCommitsPerDay: LogFields[] = []
+    for (const c of allCommits.all) {
+      const commitDate = moment(c.date).startOf('day').toString()
+      if (!allDates.includes(commitDate)) {
+        allDates.push(commitDate)
+        allLastCommitsPerDay.push(c)
+      }
+    }
+
+    // debug(`allLastCommitsPerDay=${JSON.stringify(allLastCommitsPerDay, null, 2)}`)
+    return allLastCommitsPerDay
   }
 
   private async GetModifiedFilesInCommit(hash: string): Promise<string[]> {
     const git = await this.git()
 
     const allFiles = await git.raw(['diff-tree', '--no-commit-id', '-r', '--name-only', hash])
-    return allFiles.split('\n').filter(f => f)
+    return allFiles.split('\n').filter(Boolean)
   }
 
-  private async GetAllLastCommitsPerDay(sinceDays: number): Promise<LogFields[]> {
-    const git = await this.git()
-
-    const options: LogOptions<LogFields> = {}
-    options.splitter = ';'
-    options.format = {
-      hash: '%h',
-      date: '%aI'
-    }
-    if (sinceDays > 0) {
-      options['--since'] = `"${moment().subtract(sinceDays, 'days')}"`
-    }
-    const allCommits = await git.log(options)
-    const allDates: string[] = []
-    const allLastCommitsPerDay: LogFields[] = []
-    allCommits.all.forEach(c => {
-      const commitDate = moment(c.date).startOf('day').toString()
-      if (allDates.indexOf(commitDate) === -1) {
-        allDates.push(commitDate)
-        allLastCommitsPerDay.push(c)
-      }
-    })
-    // debug(`allLastCommitsPerDay=${JSON.stringify(allLastCommitsPerDay, null, 2)}`)
-    return allLastCommitsPerDay
-  }
-
-  private async git(): Promise<simplegit.SimpleGit> {
+  private async git(): Promise<SimpleGit> {
     debug('getting git instance')
     if (!this._gitConfigAdded) {
       debug('adding config to git')
@@ -250,16 +251,18 @@ export class GitUtils {
       if (quotepath !== 'off') {
         await this._git.addConfig('core.quotepath', 'off')
       }
+
       await this._git.addConfig('user.name', this.softConfig.config.projectAuthor.name)
       await this._git.addConfig('user.email', this.softConfig.config.projectAuthor.email)
 
       this._gitConfigAdded = true
     }
+
     return this._git
   }
 }
 
 interface LogFields {
-  hash: string
   date: string
+  hash: string
 }
