@@ -1,111 +1,104 @@
-import Command, { flags } from '@oclif/command'
-import * as deb from 'debug'
+import { Args, Command, Flags, Interfaces } from '@oclif/core'
+import * as path from 'node:path'
 import * as notifier from 'node-notifier'
-import * as path from 'path'
-// import * as simplegit from 'simple-git/promise'
+import { Container, ObjectFactory, Scope } from 'typescript-ioc'
 
-import { FsUtils } from '../fs-utils'
-import { HardConfig } from '../hard-config'
-// import { promisify } from 'util'
-import '../ui-utils'
-import { Container, Scope, ObjectFactory } from 'typescript-ioc'
+import { FsUtils } from '../shared/fs-utils'
+import { HardConfig } from '../shared/hard-config'
+import { SoftConfig } from '../shared/soft-config'
+import { simpleGit } from 'simple-git'
+import { CoreUtils } from '../shared/core-utils'
+import { BootstrapChptr } from '../shared/bootstrap-functions'
+import WatchConfig from '../shared/watch-config'
 
-export const d = (cmdName: string) => {
-  return deb(`chptr:${cmdName}`)
-}
-
-// const chalk: any = require('chalk')
-// String.prototype.color = function(colorName: string) {
-//   return chalk[colorName](this)
-// }
-// String.prototype.actionStartColor = function() {
-//   return chalk.blue(this)
-// }
-// String.prototype.actionStopColor = function() {
-//   return chalk.cyan(this)
-// }
-// String.prototype.resultHighlighColor = function() {
-//   return chalk.yellow(this)
-// }
-// String.prototype.resultSecondaryColor = function() {
-//   return chalk.magenta(this)
-// }
-// String.prototype.resultNormalColor = function() {
-//   return chalk.whiteBright(this)
-// }
-// String.prototype.infoColor = function() {
-//   return chalk.gray(this)
-// }
-// String.prototype.errorColor = function() {
-//   return chalk.redBright(this)
-// }
+export const d = (cmdName: string) => require('debug')(`chptr:${cmdName}`)
 
 const debug = d('command:base')
 
-export default abstract class extends Command {
-  public get hardConfig(): HardConfig {
-    return this._hardConfig as HardConfig
-  }
-  public get fsUtils(): FsUtils {
-    return this._fsUtils as FsUtils
-  }
+export type Flags<T extends typeof Command> = Interfaces.InferredFlags<(typeof BaseCommand)['baseFlags'] & T['flags']>
+export type Args<T extends typeof Command> = Interfaces.InferredArgs<T['args']>
 
-  static flags = {
-    help: flags.help({ char: 'h' }),
-    notify: flags.boolean({
+export default abstract class BaseCommand<T extends typeof Command> extends Command {
+  static baseFlags = {
+    help: Flags.help({ char: 'h' }),
+    notify: Flags.boolean({
       char: 'N',
-      description: 'show a notification box when command is completed.',
-      default: false
+      default: false,
+      description: 'show a notification box when command is completed.'
     }),
-    path: flags.string({
+    path: Flags.string({
       char: 'p',
       default: '.',
       description: 'Path where root of project files are'
     })
   }
+
+  protected flags!: Flags<T>
+  protected args!: Args<T>
+
   static hidden = true
 
-  public get rootPath(): string {
-    return this._rootPath
-  }
-
-  private _rootPath = '.'
-  // private _git: simplegit.SimpleGit | undefined
-  private _hardConfig: HardConfig | undefined
-  private _fsUtils: FsUtils | undefined
-
-  async init() {
-    debug('Base init')
-    const { flags } = this.parse(this.constructor as any) as any
-    const dir = path.join(flags.path as string)
-    this._rootPath = dir
-
-    const hardConfigFactory: ObjectFactory = () => {
-      return new HardConfig(dir)
-    }
-
-    Container.bind(HardConfig).factory(hardConfigFactory).scope(Scope.Singleton)
-
-    this._hardConfig = Container.get(HardConfig) //new HardConfig(dir)
-
-    this._fsUtils = new FsUtils()
-  }
-
   async catch(err: Error) {
-    // throw new ChptrError(err.toString().errorColor())
-    this.error(err.toString())
-    this.exit(1)
+    this.error(err.toString(), {exit: 1})
   }
 
-  async finally() {
+  async finally(_: Error | undefined) {
     debug(`Base Finally`)
-    const { flags } = this.parse(this.constructor as any) as any
-    if (flags.notify) {
+    // const { flags } = this.parse(this.constructor as any) as any
+    if (this.flags.notify) {
       notifier.notify({
-        title: 'Chptr',
         message: `Task completed for command ${this.constructor.toString()}`,
-        sound: true
+        sound: true,
+        title: 'Chptr'
       })
     }
+
+    if (this.flags.compact) {
+      const coreUtils = Container.get(CoreUtils)
+      await coreUtils.compactFileNumbers()
+      await coreUtils.preProcessAndCommitFiles('Compacted file numbers')
+    }
+    return super.finally(_)
+  }
+
+  async init(): Promise<void> {
+    debug('Base init')
+
+    const { args, flags } = await this.parse({
+      flags: this.ctor.flags,
+      baseFlags: (super.ctor as typeof BaseCommand).baseFlags,
+      args: this.ctor.args,
+      strict: this.ctor.strict
+    })
+    this.flags = flags as Flags<T>
+    this.args = args as Args<T>
+
+    const rootPath = path.join(flags.path as string)
+
+    Container.bindName('rootPath').to(rootPath)
+
+    const softConfigFactory: ObjectFactory = () => new SoftConfig()
+    Container.bind(SoftConfig).factory(softConfigFactory).scope(Scope.Singleton)
+
+    const hardConfigFactory: ObjectFactory = () => new HardConfig(rootPath)
+    Container.bind(HardConfig).factory(hardConfigFactory).scope(Scope.Singleton)
+
+    Container.bindName('git').to(simpleGit(rootPath))
+
+    const bootstrapper = Container.get(BootstrapChptr)
+    const isChptrFolder = await bootstrapper.isChptrFolder()
+
+    if (isChptrFolder) {
+      const watchConfig = Container.get(WatchConfig)
+      await watchConfig.RenameFilesIfNewPattern()
+      await watchConfig.MoveToNewBuildDirectory()
+      await watchConfig.RenameProjectTitle()
+      await watchConfig.CheckIfStepOrInitialNumberHaveChanged()
+
+      const fsUtils = Container.get(FsUtils)
+      await fsUtils.deleteEmptySubDirectories(rootPath)
+    }
+
+    await super.init()
   }
 }
