@@ -1,55 +1,59 @@
-import { flags } from '@oclif/command'
-// import { cli } from 'cli-ux'
+import { Args, Flags } from '@oclif/core'
 
-import { ChptrError } from '../chptr-error'
-import { QueryBuilder } from '../ui-utils'
-
-import { d } from './base'
-import Command from './compactable-base'
+import { ChptrError } from '../shared/chptr-error'
+import { QueryBuilder } from '../shared/ui-utils'
+import BaseCommand, { d } from './base'
+import { compact } from '../flags/compact-flag'
+import { Container } from 'typescript-ioc'
+import { CoreUtils } from '../shared/core-utils'
+// import Command from './compactable-base'
 
 const debug = d('delete')
 
-export default class Delete extends Command {
-  static description = 'Delete a chapter or tracked file locally and in the repository'
+export default class Delete extends BaseCommand<typeof Delete> {
+  static aliases = ['del']
 
-  static flags = {
-    ...Command.flags,
-    compact: flags.boolean({
-      char: 'c',
-      description: 'Compact chapter numbers at the same time',
-      default: false
-    }),
-    save: flags.boolean({
-      char: 's',
-      description: 'Commit to git at the same time.',
-      default: false
+  static args = {
+    name: Args.string({
+      default: '',
+      description: 'chapter number or filename to delete',
+      name: 'name',
+      required: false
     })
   }
 
-  static args = [
-    {
-      name: 'name',
-      description: 'chapter number or filename to delete',
-      required: false,
-      default: ''
-    }
-  ]
+  static description = 'Delete a chapter or tracked file locally and in the repository'
 
-  static aliases = ['del']
+  static flags = {
+    compact: compact,
+    // compact: Flags.boolean({
+    //   char: 'c',
+    //   default: false,
+    //   description: 'Compact chapter numbers at the same time'
+    // }),
+    save: Flags.boolean({
+      char: 's',
+      default: false,
+      description: 'Commit to git at the same time.'
+    })
+  }
 
   static hidden = false
 
   async run() {
     debug('Running Delete command')
-    const { args, flags } = this.parse(Delete)
+    const { args, flags } = await this.parse(Delete)
 
-    const compact = flags.compact
-    const save = flags.save
+    const { compact } = flags
+    const { save } = flags
+
+    const coreUtils = Container.get(CoreUtils)
 
     const queryBuilder = new QueryBuilder()
     if (!args.name) {
       queryBuilder.add('name', queryBuilder.textinput('Chapter number or filename to delete?'))
     }
+
     const queryResponses: any = await queryBuilder.responses()
     const nameOrNumber: any = args.name || queryResponses.name
 
@@ -57,16 +61,15 @@ export default class Delete extends Command {
       throw new ChptrError('Name or number input empty', 'delete.run', 4)
     }
 
-    let commitMsg = await this.coreUtils.deleteFilesFromRepo(nameOrNumber)
+    let commitMsg = await coreUtils.deleteFilesFromRepo(nameOrNumber)
 
     if (compact) {
-      await this.coreUtils.compactFileNumbers()
-      // await this.gitWrapper.CommitToGit(`Compacted file numbers`)
+      await coreUtils.compactFileNumbers()
       commitMsg += `\nCompacted file numbers`
     }
 
     if (save) {
-      await this.coreUtils.preProcessAndCommitFiles(commitMsg, undefined, true)
+      await coreUtils.preProcessAndCommitFiles(commitMsg, undefined, true)
     }
   }
 }
