@@ -1,56 +1,64 @@
-import { flags } from '@oclif/command'
-import { cli } from 'cli-ux'
-import * as path from 'path'
+import { Args, Flags, ux } from '@oclif/core'
+import { glob } from 'glob'
+import * as path from 'node:path'
 
-import { ChapterId } from '../chapter-id'
-import { ChptrError } from '../chptr-error'
-import { QueryBuilder } from '../ui-utils'
-
-import { d } from './base'
-import Command from './initialized-base'
+import { ChapterId } from '../shared/chapter-id'
+import { ChptrError } from '../shared/chptr-error'
+import { QueryBuilder } from '../shared/ui-utils'
+import BaseCommand, { d } from './base'
+import { Container } from 'typescript-ioc'
+import { SoftConfig } from '../shared/soft-config'
+import { FsUtils } from '../shared/fs-utils'
+import { CoreUtils } from '../shared/core-utils'
+// import Command from './initialized-base'
 
 const debug = d('edit')
 
-export default class Edit extends Command {
+export default class Edit extends BaseCommand<typeof Edit> {
+  static aliases = ['modify', 'mod']
+
+  static args = {
+    chapterIds: Args.string({
+      default: '',
+      description: 'Chapter number(s) to modify, comma-separated or dash-separated for a range.',
+      name: 'chapterIds',
+      required: false
+    })
+  }
+
   static description = 'Adjust sentence and paragraph endings to allow for easier editing.  Commit changes with SAVE command.'
 
   static flags = {
-    ...Command.flags,
-    type: flags.string({
+    type: Flags.string({
       char: 't',
-      description: 'Edit either chapter file, summary file or all.',
       default: 'all',
+      description: 'Edit either chapter file, summary file or all.',
       options: ['all', 'summary', 'chapter']
     })
   }
 
-  static args = [
-    {
-      name: 'chapterIds',
-      description: 'Chapter number(s) to modify, comma-separated or dash-separated for a range.',
-      required: false,
-      default: ''
-    }
-  ]
-
-  static aliases = ['modify', 'mod']
+  static hidden = false
 
   // for variable length arguments (https://oclif.io/docs/args)
   static strict = false
 
-  static hidden = false
-
   async run() {
     debug('Running Edit command')
-    const { flags, argv } = this.parse(Edit)
+
+    const softConfig = Container.get(SoftConfig)
+    const rootPath = Container.getValue('rootPath')
+    const fsUtils = Container.get(FsUtils)
+    const coreUtils = Container.get(CoreUtils)
+
+    const { argv, flags } = await this.parse(Edit)
 
     const editType = flags.type
 
-    const argsArray: string[] = argv
+    const argsArray: string[] = argv as string[]
     const chapterIdsString: string[] = []
     const toEditFiles: string[] = []
     if (!argsArray[0]) {
-      //no chapter given; must ask for it
+      // no chapter given; must ask for it
       const queryBuilder = new QueryBuilder()
       queryBuilder.add(
         'chapterIds',
@@ -61,15 +69,18 @@ export default class Edit extends Command {
       argsArray.push(...queryResponses.chapterIds.split(' '))
     }
 
-    //loop through all argv[i] to get all chapter numbers.  If any argument contains commas, it will be split at ','.  If it has dashes, it will be interpolated.
+    // loop through all argv[i] to get all chapter numbers.  If any argument contains commas, it will be split at ','.  If it has dashes, it will be interpolated.
     for (const arg of argsArray) {
       if (arg.split(',').length > 1) {
         chapterIdsString.push(...arg.split(','))
       } else if (arg.split('-').length == 2) {
-        const lowRange: number = parseInt(arg.split('-')[0], 10)
-        const hiRange: number = parseInt(arg.split('-')[1], 10)
+        const lowRange: number = Number.parseInt(arg.split('-')[0], 10)
+        const hiRange: number = Number.parseInt(arg.split('-')[1], 10)
         // const range = Array.from({ length: hiRange - lowRange }, (v, k) => k + 1).map(v => v.toString())
-        const range = Array(hiRange - lowRange + 1).fill(undefined, undefined, undefined).map((_, idx) => lowRange + idx).map(v => v.toString())
+        const range = new Array(hiRange - lowRange + 1)
+          .fill(0)
+          .map((_, idx) => lowRange + idx)
+          .map(v => v.toString())
         debug(`range: ${JSON.stringify(range)}`)
         chapterIdsString.push(...range)
       } else {
@@ -78,13 +89,14 @@ export default class Edit extends Command {
     }
 
     for (const id of chapterIdsString.map(
-      (input: string) => new ChapterId(this.softConfig.extractNumber(input), this.softConfig.isAtNumbering(input))
+      (input: string) => new ChapterId(softConfig.extractNumber(input), softConfig.isAtNumbering(input))
     )) {
       if (editType === 'all' || editType === 'chapter') {
-        toEditFiles.push(...(await this.fsUtils.listFiles(path.join(this.rootPath, this.softConfig.chapterWildcardWithNumber(id)))))
+        toEditFiles.push(...(await glob(path.join(rootPath, softConfig.chapterWildcardWithNumber(id)))))
       }
+
       if (editType === 'all' || editType === 'summary') {
-        toEditFiles.push(...(await this.fsUtils.listFiles(path.join(this.rootPath, this.softConfig.summaryWildcardWithNumber(id)))))
+        toEditFiles.push(...(await glob(path.join(rootPath, softConfig.summaryWildcardWithNumber(id)))))
       }
     }
 
@@ -92,15 +104,16 @@ export default class Edit extends Command {
       throw new ChptrError('No files matching input', 'edit.run', 20)
     }
 
-    cli.action.start('Reading and processing files'.actionStartColor())
+    ux.action.start('Reading and processing files'.actionStartColor())
     for (const filename of toEditFiles) {
-      const fullPath = path.join(this.rootPath, filename)
+      const fullPath = path.join(rootPath, filename)
 
-      const initialContent = await this.fsUtils.readFileContent(fullPath)
-      const replacedContent = await this.coreUtils.processContentBack(initialContent)
-      await this.fsUtils.writeFile(fullPath, replacedContent)
+      const initialContent = await fsUtils.readFileContent(fullPath)
+      const replacedContent = await coreUtils.processContentBack(initialContent)
+      await fsUtils.writeFile(fullPath, replacedContent)
     }
+
     const toEditPretty = toEditFiles.map(f => `\n    ${f}`)
-    cli.action.stop(`${toEditPretty}\n${toEditFiles.length > 1 ? 'were' : 'was'} modified`.actionStopColor())
+    ux.action.stop(`${toEditPretty}\n${toEditFiles.length > 1 ? 'were' : 'was'} modified`.actionStopColor())
   }
 }
