@@ -1,57 +1,44 @@
-import * as simplegit from 'simple-git/promise'
+import { SimpleGit } from 'simple-git'
+import { Inject, InjectValue, Singleton } from 'typescript-ioc'
 
-import { d } from './commands/base'
 import { FsUtils } from './fs-utils'
 import { HardConfig } from './hard-config'
-import { Singleton, Container } from 'typescript-ioc'
 
-const debug = d('BootstrapChptr')
+const debug = require('debug')('BootstrapChptr')
 
 @Singleton
 export class BootstrapChptr {
-  private readonly rootPath: string
-  private readonly hardConfig: HardConfig
+  private readonly git: SimpleGit
   private readonly fsUtils: FsUtils
+  private readonly hardConfig: HardConfig
 
-  constructor(rootPath: string) {
-    this.rootPath = rootPath
-    this.hardConfig = Container.get(HardConfig)//new HardConfig(rootPath)
-    this.fsUtils = new FsUtils()
-  }
-
-  public async isChptrFolder(): Promise<boolean> {
-    const git = simplegit(this.rootPath)
-    const isRepo = await git.checkIsRepo()
-    const hasConfigFolder = await this.fsUtils.fileExists(this.hardConfig.configPath)
-    const hasConfigJSON5File = await this.fsUtils.fileExists(this.hardConfig.configJSON5FilePath)
-    const hasConfigYAMLFile = await this.fsUtils.fileExists(this.hardConfig.configYAMLFilePath)
-
-    if (!isRepo || !hasConfigFolder || !(hasConfigJSON5File || hasConfigYAMLFile)) {
-      return false
-    } else {
-      return true
-    }
+  constructor(@InjectValue('git') git: SimpleGit, @Inject hardConfig: HardConfig, @Inject fsUtils: FsUtils) {
+    this.git = git
+    this.hardConfig = hardConfig
+    this.fsUtils = fsUtils
   }
 
   public deepCopy(obj: any): any {
     return Object.keys(obj).reduce((retval: any, key: string) => {
       if (obj[key]) {
         if (typeof obj[key] !== 'string' && typeof obj[key] !== 'number' && typeof obj[key] !== 'boolean') {
-          if (obj[key] instanceof Array && obj[key].length === 0) {
+          if (Array.isArray(obj[key]) && obj[key].length === 0) {
             return retval
           }
+
           if (Object.keys(obj[key]).length === 0) {
             return retval
           }
 
           const deepObj = this.deepCopy(obj[key])
-          if (Object.keys(deepObj).length !== 0) {
+          if (Object.keys(deepObj).length > 0) {
             retval[key] = deepObj
           }
         } else {
           retval[key] = obj[key]
         }
       }
+
       return retval
     }, {})
     //  https://stackoverflow.com/questions/28150967/typescript-cloning-object/42758108
@@ -106,5 +93,18 @@ export class BootstrapChptr {
 
     throw new Error("Unable to copy obj! Its type isn't supported.")
     */
+  }
+
+  public async isChptrFolder(): Promise<boolean> {
+    const isRepo = await this.git.checkIsRepo()
+    const hasConfigFolder = await this.fsUtils.fileExists(this.hardConfig.configPath)
+    const hasConfigJSON5File = await this.fsUtils.fileExists(this.hardConfig.configJSON5FilePath)
+    const hasConfigYAMLFile = await this.fsUtils.fileExists(this.hardConfig.configYAMLFilePath)
+
+    if (!isRepo || !hasConfigFolder || !(hasConfigJSON5File || hasConfigYAMLFile)) {
+      return false
+    }
+
+    return true
   }
 }
